Completed the database work
Relations are working by adding the appropriate object prop
called associate (function) and passing in the models
Added a flag for dev mode, which will recreate the database on every
start of the application.

There is currently a problem with model.findall|count where bluebird
will print a warning complaining about a unhandled promise.  As far
as I can tell that is caused by Sequelize and not me.